requirements.txt
aiofiles==0.7.0
aiojobs==0.3.0
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.16.1
anyio==3.3.4
async-timeout==3.0.1
asyncpg==0.25.0
atlastk==0.13
Babel==2.9.1
bcrypt==3.2.0
bleach==3.3.0
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
ddlparse==1.10.0
dictdiffer==0.9.0
ecdsa==0.17.0
Faker==8.8.1
fastapi==0.75.2
filetype==1.0.7
furl==2.1.3
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.3
install==1.3.4
iso8601==1.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
orderedmultidict==1.0.1
orjson==3.6.3
packaging==20.9
passlib==1.7.4
pendulum==2.1.2
pubg-python==0.12.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.4
python-dateutil==2.8.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2020.5
pytzdata==2020.1
PyYAML==6.0
requests==2.27.1
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sse-starlette==0.10.0
starlette==0.17.1
text-unidecode==1.3
tomlkit==0.8.0
tortoise-orm==0.19.0
typing-extensions==3.7.4.3
urllib3==1.26.9
uvicorn==0.13.4
watchgod==0.7
webencodings==0.5.1
websockets==8.1
